Saved with strategy pattern for product creation

+++ CleanBrilliantCompany/Interfaces/iProductQuery.cs
@@ -17,7 +17,7 @@ namespace CleanBrilliantCompany.Interfaces
         List<ProductBatch> getAllProductBatch();
         ProductBatch getBatchDetails(int batchCode); 
         public int createProductBatch(int productId, DateTime expiryDate, 
-            DateTime receiveDate, DateTime manufactureDate, int quantity, int batchCost);
+            DateTime receiveDate, DateTime manufactureDate, int quantity, float batchCost);
 
         // Stock History
         // List<StockHistory> getStockHistoryByBatch(int batchCode);
